pyproject.toml
[project]
name = "saas-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "fastapi[standard]>=0.135.2",
    "uvicorn>=0.30.0",
    "pydantic>=2.7.0",
    "pre-commit>=4.5.1",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
    "sqlalchemy[asyncio]>=2.0.48",
    "asyncpg>=0.31.0",
    "pydantic-settings>=2.13.1",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.12.1",
]

[project.optional-dependencies]
dev = [
    "black>=26.3.1",
    "ruff>=0.15.7",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.13.1",
]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"
fix = true

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
strict = true
disallow_untyped_defs = true
disallow_untyped_decorators = false
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_unused_ignores = true
warn_return_any = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]
